Conflicts:
	3rdparty/pythia8215/bin/pythia8-config
	src/experiment.cc
	src/include/modusdefault.h
All solved

+++ src/nucleus.cc
@@ -39,8 +39,8 @@ Nucleus::Nucleus(Configuration &config, int nTest) {
   }
 }
 
-float Nucleus::mass() const {
-  float total_mass = 0.f;
+double Nucleus::mass() const {
+  double total_mass = 0.;
   for (auto i = cbegin(); i != cend(); i++) {
     total_mass += i->momentum().abs();
   }
